Reinicio orgrnizaacion

+++ app/Http/Controllers/CustomerController.php
@@ -18,7 +18,7 @@ class CustomerController extends Controller
         $validated = $request->validate([
             'nombre' => "required|max:255",
             'direccion' => "required|max:255",
-            'cif' => "required|max:9",
+            'cif' => "required|min:9|max:9",
             'mail' => "required|max:100",
             'telefono' => "required|max:9",
             'encargado' => "required|max:100",
@@ -29,9 +29,11 @@ class CustomerController extends Controller
 
         Customer::create($validated);
 
+
         //return dd($customer);
 
-        return redirect (route('list-customer'));
+        //return redirect (route('list-customer'));
+        return redirect('/app/customer-list');
     }
 
     public function show($slug)
@@ -40,15 +42,19 @@ class CustomerController extends Controller
         return view('components.customer.customer', ["customer"=>$customer]);
     }
 
-    public function destroy(Customer $customer)
+    public function delete(Request $request)
     {
-        $customer->delete();
-        return redirect()->view('components.customer.list')
-            ->with('succes', 'Cliente eliminado correctamente');
+        $id = $request->input('customer_id');
+        $customer = Customer::findOrFail($id);
+        if($customer->user == auth()->user()){
+            $customer->delete();
+        }
+        return redirect('/app/customer-list')->with('success', 'Cliente eliminado correctamente');
     }
 
     public function list()
     {
-        return view('components.customer.list', ["customers" => Customer::all()]);
+        //return view('components.customer.list', ["customers" => Customer::all()]);
+        return view('admin.customers', ["customers" => Customer::all()]);
     }
 }
